federated_ckd/central_server/requirements.txt
psycopg2-binary==2.9.10
numpy==1.25.2
